backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic[email]==2.9.2
httpx==0.27.2
python-multipart==0.0.12
pymongo==4.8.0
python-dotenv==1.0.1
slowapi==0.1.9

# MCP and ML dependencies
torch>=2.0.0
setfit>=1.0.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0
mlflow>=2.0.0
google-cloud-storage>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0

# MongoDB dependencies
pymongo[srv]==4.8.0
bson==0.5.10
